# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Piano Shop Api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/user_show_by_id/{id}": {
            "get": {
                "tags": [
                    "User Api"
                ],
                "description": "use for get user data",
                "operationId": "6db27b75c7df53ab34917f1c17f53065",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/change_password/{id}": {
            "post": {
                "tags": [
                    "User Api"
                ],
                "description": "use for change password",
                "operationId": "01d93978610fe8b80a4e296eb5f7060a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "description": "Enter your password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/user_delete/{id}": {
            "delete": {
                "tags": [
                    "User Api"
                ],
                "description": "use for delete user",
                "operationId": "df0912d8678637b1922c0d172ccdd285",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/user_update/{id}": {
            "put": {
                "tags": [
                    "User Api"
                ],
                "description": "use for update user",
                "operationId": "bd6410c4c972ba374e8df7b8fa6a0894",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Enter your name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Enter your email",
                                        "type": "string"
                                    },
                                    "role": {
                                        "description": "Enter your role",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Enter your phone",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "Enter your address",
                                        "type": "string"
                                    },
                                    "country": {
                                        "description": "Enter your country",
                                        "type": "string"
                                    },
                                    "city": {
                                        "description": "Enter your city",
                                        "type": "string"
                                    },
                                    "postCode": {
                                        "description": "Enter your postCode",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/v1/payment": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Create a new payment",
                "operationId": "44472dc218d6766a42d217b87c4c4c14",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "pi_1J2K3K4L5M6N7O8P9Q0R"
                                        },
                                        "object": {
                                            "type": "string",
                                            "example": "payment_intent"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "succeeded"
                                        },
                                        "amount": {
                                            "type": "integer",
                                            "example": 1000
                                        },
                                        "currency": {
                                            "type": "string",
                                            "example": "gbp"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Payment not created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment Not Created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/payment_confirm": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Confirm a payment",
                "operationId": "3cef79d64c2421e752cf7d1d849d59f7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment_intent_id": {
                                        "type": "string",
                                        "example": "pi_1J2K3K4L5M6N7O8P9Q0R"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment confirmed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment confirmed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Payment not confirmed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Payment Not Confirmed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}